+++ .env
@@ -1,0 +1,1 @@
+VITE_APP_BASE_URL=/api
