+++ package.json
@@ -16,6 +16,7 @@
     "antd": "^5.5.1",
     "axios": "^1.4.0",
     "dayjs": "^1.11.7",
+    "dotenv": "^16.1.4",
     "graphql": "^16.6.0",
     "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
