+++ public/_redirects
@@ -1,0 +1,2 @@
+/api/*   http://testing.biaracmpny.my.id/:splat 200
+/*  /index.html 200
