+++ src/api/Index.jsx
@@ -1,6 +1,24 @@
-import { newBaseAPI, baseAPI } from "../config/apiService";
+import { baseAPI } from "../config/apiService";
+
+baseAPI.interceptors.request.use(
+  (config) => {
+    const token = localStorage.getItem("token");
+    if (token) {
+      config.headers.Authorization = `Bearer ${token}`;
+    }
+    return config;
+  },
+  (error) => {
+    return Promise.reject(error);
+  }
+);
 
 export const api = {
+  /// API with Token
+  // Login
+  login: (body) => {
+    return baseAPI.post("/v1/users/login", body);
+  },
   getUsers: () => {
     return baseAPI.get("/v1/users");
   },
@@ -10,9 +28,16 @@ export const api = {
   getMahasiswa: () => {
     return baseAPI.get("/v1/mahasiswa");
   },
-  /// API with Token
-  // Login
-  login: (body) => {
-    return newBaseAPI.post("/v1/users/login", body);
+  getDosenById: (id) => {
+    return baseAPI.get(`/v1/dosen/${id}`);
+  },
+  signUp: (body) => {
+    return baseAPI.post("/v1/users/signup", body);
+  },
+  createDosen: (body) => {
+    return baseAPI.post("/v1/dosen", body);
+  },
+  createMahasiswa: (body) => {
+    return baseAPI.post("/v1/mahasiswa", body);
   },
 };
