+++ src/config/apiService.jsx
@@ -1,52 +1,9 @@
 import axios from "axios";
+import { BASE_URL } from "../utils";
 
-const baseAPI = axios.create({
-  baseURL: "http://testing.biaracmpny.my.id",
+export const baseAPI = axios.create({
+  baseURL: BASE_URL.API,
+  headers: {
+    "Content-Type": "application/json",
+  },
 });
-
-// Menambahkan interceptor untuk menyertakan token pada setiap permintaan
-baseAPI.interceptors.request.use(
-  (config) => {
-    const token = localStorage.getItem("token");
-    if (token) {
-      config.headers.Authorization = `Bearer ${token}`;
-    }
-    return config;
-  },
-  (error) => {
-    return Promise.reject(error);
-  }
-);
-
-const api = {
-  login: (credentials) => {
-    return baseAPI.post("/v1/users/login", credentials);
-  },
-  getUsers: () => {
-    return baseAPI.get("/v1/users");
-  },
-  getDosen: () => {
-    return baseAPI.get("/v1/dosen");
-  },
-  getMahasiswa: () => {
-    return baseAPI.get("/v1/mahasiswa");
-  },
-  getDosenById: (id) => {
-    return baseAPI.get(`/v1/dosen/${id}`);
-  },
-  signUp: (body) => {
-    return baseAPI.post("/v1/users/signup", body);
-  },
-  createDosen: (body) => {
-    return baseAPI.post("/v1/dosen", body);
-  },
-  createMahasiswa: (body) => {
-    return baseAPI.post("/v1/mahasiswa", body);
-  },
-};
-
-const newBaseAPI = axios.create({
-  baseURL: "http://testing.biaracmpny.my.id",
-});
-
-export { baseAPI, api, newBaseAPI };
