+++ src/utils/index.jsx
@@ -1,0 +1,3 @@
+export const BASE_URL = {
+  API: `${import.meta.env.VITE_APP_BASE_URL}`,
+};
